# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.RegisterRequest:
    properties:
      cityId:
        type: integer
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      role:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - role
    type: object
...
  models.Building:
    properties:
      address:
        type: string
      category:
        $ref: '#/definitions/models.BuildingCategory'
      category_id:
        type: integer
      city:
        $ref: '#/definitions/models.City'
      city_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      orders:
        items:
          $ref: '#/definitions/models.ReconstructionOrder'
        type: array
      resources:
        items:
          $ref: '#/definitions/models.BuildingResource'
        type: array
    type: object
  models.BuildingCategory:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.BuildingResource:
    properties:
      building_id:
        type: integer
      id:
        type: integer
      is_main:
        type: boolean
      resource_type:
        type: string
      url:
        type: string
    type: object
  models.City:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Donation:
    properties:
      amount:
        type: number
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      order_id:
        type: integer
      user_id:
        type: integer
    type: object
  models.OrderService:
    properties:
      description:
        type: string
      id:
        type: integer
      order_id:
        type: integer
      price:
        type: number
      service:
        $ref: '#/definitions/models.Service'
      service_id:
        type: integer
    type: object
  models.ReconstructionOrder:
    properties:
      building:
        $ref: '#/definitions/models.Building'
      building_id:
        type: integer
      collected_amount:
        type: number
      completed_at:
        type: string
      created_at:
        type: string
      creator:
        $ref: '#/definitions/models.User'
      creator_id:
        type: integer
      donations:
        items:
          $ref: '#/definitions/models.Donation'
        type: array
      id:
        type: integer
      moderator_id:
        type: integer
      services:
        items:
          $ref: '#/definitions/models.OrderService'
        type: array
      status:
        type: string
      total_amount:
        type: number
    type: object
  models.Service:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      status:
        type: string
      video_url:
        type: string
    type: object
  models.User:
    properties:
      city:
        $ref: '#/definitions/models.City'
      city_id:
        type: integer
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API documentation for History Care
  title: History Care API
  version: "1.0"
paths:
  /api/auth/cities:
    get:
      description: Возвращает список городов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.City'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all cities
      tags:
      - auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Логин пользователя, возвращает JWT
      parameters:
      - description: Login credentials
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
...
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя и возвращает JWT
      parameters:
      - description: User registration data
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /api/buildings:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Building name
        in: formData
        name: name
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Address
        in: formData
        name: address
        required: true
        type: string
      - description: Category ID
        in: formData
        name: category_id
        required: true
        type: integer
      - description: City ID (only for admin)
        in: formData
        name: city_id
        type: integer
      - collectionFormat: csv
        description: Photos + Videos
        in: formData
        items:
          type: file
        name: files
        required: true
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Building'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create building
      tags:
      - buildings
  /api/buildings/{id}:
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Building ID
        in: path
        name: id
        required: true
        type: integer
      - description: Building name
        in: formData
        name: name
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Address
        in: formData
        name: address
        type: string
      - description: Category ID
        in: formData
        name: category_id
        type: integer
      - description: City ID
        in: formData
        name: city_id
        type: integer
      - collectionFormat: csv
        description: Photos + Videos
        in: formData
        items:
          type: file
        name: files
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Building'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update building
      tags:
      - buildings
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"